fix bug

+++ commands/owner only/setmoney.js
@@ -17,24 +17,33 @@ module.exports = {
      * @returns 
      */
     async execute(client, message, args) {
-        let userId = getUserId(client,message,args,false);
+        var user = args[0] || message.author.id;
+        var tag = message.mentions.members.first();
 
-        trys(userId);
-        function trys(userId) {
-            client.users.fetch(userId).then(async user => {
-                if(!args[1] || isNaN(args[1])) return message.reply({ content: "Nhập số hợp lệ", allowedMentions: { repliedUser: false } })
-                    .then(msg => client.deleteMsg(msg));
-                
-                setMoney(message.author.tag, user.user.id, toCheck);
+        if (isNaN(user) && !tag) user = message.author.id;
+        if (tag) user = tag.id;
 
-                message.reply({ embeds:[{
-                    description:"Tiền của **" + user.user.tag + "** đã được đặt thành " + sodep(toCheck) + " " + client.emoji.dongxu,
-                    color: client.config.DEF_COLOR
-                }], allowedMentions: { repliedUser: false } });
-            }).catch(err => {
-                if(err.message == "Unknown User") return message.reply({contnet: "Người nhận không hợp lệ", allowedMentions: { repliedUser: false} });
-                message.botError();
-            });
-        }
+        let check_name = client.users.cache.find(user => user.username.toLowerCase() == args.join(" ").toLowerCase());
+        if (check_name) user = check_name.id;
+
+        if (!check_name && !tag && isNaN(user) || !tag && isNaN(user)) return message.reply({
+            content: "Nhập cấp người nhận", allowedMentions: { repliedUser: false }
+        }).then(msg => client.msgDelete(msg));
+
+        client.users.fetch(user).then(async user => {
+            if(!args[1] || isNaN(args[1])) return message.reply({ content: "Nhập số hợp lệ", allowedMentions: { repliedUser: false } })
+                .then(msg => client.deleteMsg(msg));
+            
+            setMoney(message.author.tag, user.user.id, toCheck);
+
+            message.reply({ embeds:[{
+                description:"Tiền của **" + user.user.tag + "** đã được đặt thành " + sodep(toCheck) + " " + client.emoji.dongxu,
+                color: client.config.DEF_COLOR
+            }], allowedMentions: { repliedUser: false } });
+        }).catch(err => {
+            if(err.message == "Unknown User") return message.reply({contnet: "Người nhận không hợp lệ", allowedMentions: { repliedUser: false} });
+            message.botError();
+            client.sendError(err.toString());
+        });
     }
 }
